готовая версия

+++ myshop/images/management/commands/fillimages.py
@@ -1,6 +1,7 @@
 from django.core.management.base import BaseCommand, CommandError
 from django.conf import settings
 import os
+from shutil import copy2
 from images.models import Image
 
 
@@ -16,17 +17,25 @@ class Command(BaseCommand):
         img_formats = ('.jpg', '.jpeg', '.png')
         imgs_dir = options['directory']
 
-        # path = os.path.join(settings.MEDIA_ROOT, imgs_dir)
         path = os.path.join(settings.BASE_DIR, imgs_dir)
         file_list = os.listdir(path)
 
         if options['del']:
-            Image.objects.all().delete()
+            confirm = input('do you want to delete all images, including in media directory? y/n ')
+            if confirm.lower() == 'y':
+                Image.objects.all().delete()
+                medai_path = os.path.join(settings.BASE_DIR, 'media')
+                media_file_list = os.listdir(medai_path)
+                for f in media_file_list:
+                    if os.path.isfile(os.path.join(medai_path, f)):
+                        os.remove(medai_path, f)
 
-        for f in file_list:
-            if f.endswith(img_formats):
-                image = Image(
-                    name=f,
-                    img=f
-                )
-                image.save()
+        for d in file_list:
+            for f in os.listdir(os.path.join(path, d)):
+                if f.endswith(img_formats):
+                    image = Image(
+                        name=d + '_' + f,
+                        img=f
+                    )
+                    image.save()
+                    copy2(os.path.join(path, d, f), os.path.join(settings.BASE_DIR, 'media'))
